Refactor rich text editor into a folder
