+++ packages/frontend/src/components/index.ts
@@ -10,5 +10,5 @@ export * from "./inline_input";
 export * from "./name_input";
 export * from "./panel";
 export * from "./resizable";
-export * from "./rich_text_editor";
+export * from "./rich_text_editor/rich_text_editor";
 export * from "./json_import";
