+++ packages/frontend/src/components/rich_text_editor/rich_text_editor.css
